requirementsA.txt
fastapi
uvicorn
pydantic
pydantic_settings
python-dotenv
dependency_injector
loguru
